create service for category

+++ src/main/java/com/joel/br/JJ/TECH/services/ProductService.java
@@ -1,11 +1,13 @@
 package com.joel.br.JJ.TECH.services;
 
+import com.joel.br.JJ.TECH.DTO.CategoryDTO;
 import com.joel.br.JJ.TECH.DTO.ProductDTO;
 import com.joel.br.JJ.TECH.exceptions.DataBaseException;
 import com.joel.br.JJ.TECH.exceptions.ProductNotFoundException;
+import com.joel.br.JJ.TECH.models.Category;
 import com.joel.br.JJ.TECH.models.Product;
+import com.joel.br.JJ.TECH.repository.CategoryRepository;
 import com.joel.br.JJ.TECH.repository.ProductRepository;
-import org.hibernate.exception.DataException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,9 +18,11 @@ public class ProductService {
 
 
     private final ProductRepository repository;
+    private final CategoryRepository categoryRepository;
 
-    public ProductService(ProductRepository repository) {
+    public ProductService(ProductRepository repository, CategoryRepository categoryRepository) {
         this.repository = repository;
+        this.categoryRepository = categoryRepository;
     }
 
 
@@ -55,6 +59,10 @@ public class ProductService {
 
     public void  deleteById(Long id) throws  Exception {
 
+        if(!repository.existsById(id)) {
+            throw  new ProductNotFoundException("product not found");
+        }
+
         try {
             repository.deleteById(id);
         }catch (DataIntegrityViolationException e){
@@ -77,9 +85,17 @@ public class ProductService {
 
 
     public void  copyEntity(Product product, ProductDTO dto) {
+        product.setId(product.getId());
         product.setName(dto.getName());
         product.setDescription(dto.getDescription());
         product.setPrice(dto.getPrice());
         product.setImgUrl(dto.getImgUrl());
+
+
+
+        for( CategoryDTO category : dto.getCategories()){
+            Category category1 = categoryRepository.getOne(category.getId());
+            product.getCategories().add(category1);
+        }
     }
 }
